[PRACTICUM REVIEW] feat(fullScreen): add fullscreen api for game (GAM-32)
