+++ packages/client/src/shared/hooks/useFullScreenAPI.tsx
@@ -1,0 +1,32 @@
+import { useEffect, useState } from 'react';
+
+// Функция для включения/выключения полноэкранного режима
+const toggleFullscreen = () => {
+    if (!document.fullscreenElement) {
+        document.documentElement.requestFullscreen();
+    } else {
+        if (document.exitFullscreen) {
+            document.exitFullscreen();
+        }
+    }
+};
+
+export const useFullScreenAPI = () => {
+    const [isFullscreen, setIsFullscreen] = useState(false);
+
+    useEffect(() => {
+        const onFullscreenChange = () => {
+            setIsFullscreen(Boolean(document.fullscreenElement));
+        };
+
+        document.addEventListener('fullscreenchange', onFullscreenChange);
+
+        return () =>
+            document.removeEventListener(
+                'fullscreenchange',
+                onFullscreenChange
+            );
+    }, []);
+
+    return { isFullscreen, toggleFullscreen };
+};
