+++ packages/client/src/shared/ui/index.ts
@@ -2,3 +2,4 @@ export * from './PageLoader';
 export * from './Container';
 export * from './AppLink';
 export * from './Card';
+export * from './Icon';
